Private function 'Set-RootModule' - rewritten and fixed.
Private function 'Set-RootModule' - rewritten to use functions 'Get-ModuleFilesIncluded' and 'Get-ModuleFilesInventory'. Fixed errata in line 60.
Now returns the root module's FileInfo even if the type of the root module is 'Manifest'.
Function 'New-ModuleManifestAuto' - now does not includes root module to manifest data if the type of it is 'Manifest'.
Function 'Get-NestedModules' - typename of the parameter 'RootModule' is 'System.IO.FileInfo'. Removed unnecessary actions such as getting the list of module files etc.